MutableTransaction should not extend Transaction
