+++ src/Transaction/Transaction.php
@@ -9,37 +9,17 @@ use BitWasp\Bitcoin\Crypto\Hash;
 use BitWasp\Bitcoin\Serializable;
 use BitWasp\Bitcoin\Serializer\Transaction\TransactionSerializer;
 
-class Transaction extends Serializable implements TransactionInterface
+class Transaction extends AbstractTransaction implements TransactionInterface
 {
     /**
-     * @var int|string
+     * @var Buffer|null
      */
-    protected $version;
-
-    /**
-     * @var TransactionInputCollection
-     */
-    protected $inputs;
-
-    /**
-     * @var TransactionOutputCollection
-     */
-    protected $outputs;
-
-    /**
-     * @var int|string
-     */
-    protected $locktime;
+    private $raw = null;
 
     /**
      * @var string|null
      */
     private $txId = null;
-
-    /**
-     * @var Buffer|null
-     */
-    private $raw = null;
 
     /**
      * @param int|string                  $version
@@ -88,82 +68,6 @@ class Transaction extends Serializable implements TransactionInterface
         return $this->txId;
     }
 
-    protected function createTransactionId() {
-        return bin2hex(Buffertools::flipBytes(Hash::sha256d($this->getBuffer())));
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function getVersion()
-    {
-        return $this->version;
-    }
-
-    /**
-     * Get the array of inputs in the transaction
-     *
-     * @return TransactionInputCollection
-     */
-    public function getInputs()
-    {
-        return $this->inputs;
-    }
-
-    /**
-     * Get Outputs
-     *
-     * @return TransactionOutputCollection
-     */
-    public function getOutputs()
-    {
-        return $this->outputs;
-    }
-
-    /**
-     * Get Lock Time
-     *
-     * @return int|string
-     */
-    public function getLockTime()
-    {
-        return $this->locktime;
-    }
-
-    /**
-     * @return SignatureHash
-     */
-    public function getSignatureHash()
-    {
-        return new SignatureHash($this);
-    }
-
-    /**
-     * @return TransactionInterface
-     */
-    public function makeImmutableCopy()
-    {
-        return new Transaction(
-            $this->getVersion(),
-            $this->getInputs()->makeImmutableCopy(),
-            $this->getOutputs()->makeImmutableCopy(),
-            $this->getLockTime()
-        );
-    }
-
-    /**
-     * @return MutableTransactionInterface
-     */
-    public function makeMutableCopy()
-    {
-        return new MutableTransaction(
-            $this->getVersion(),
-            $this->getInputs()->makeMutableCopy(),
-            $this->getOutputs()->makeMutableCopy(),
-            $this->getLockTime()
-        );
-    }
-
     /**
      * @return Buffer
      */
@@ -174,13 +78,5 @@ class Transaction extends Serializable implements TransactionInterface
         }
 
         return $this->raw;
-    }
-
-    protected function createBuffer()
-    {
-        $serializer = new TransactionSerializer();
-        $raw = $serializer->serialize($this);
-
-        return $raw;
     }
 }
